:art: Refactor file structure to support discussions
