+++ components/consistent-evaluation-page.js
@@ -1,8 +1,8 @@
 import './header/d2l-consistent-evaluation-learner-context-bar.js';
-import './left-panel/consistent-evaluation-left-panel.js';
+import './left-panel/assignments/consistent-evaluation-left-panel.js';
 import './footer/consistent-evaluation-footer-presentational.js';
 import './right-panel/consistent-evaluation-right-panel.js';
-import './left-panel/consistent-evaluation-submissions-page.js';
+import './left-panel/assignments/consistent-evaluation-submissions-page.js';
 import './header/consistent-evaluation-nav-bar.js';
 import './consistent-evaluation-dialogs.js';
 import '@brightspace-ui/core/components/alert/alert-toast.js';
