+++ components/left-panel/assignments/consistent-evaluation-evidence-file.js
@@ -1,8 +1,8 @@
 import './consistent-evaluation-evidence-top-bar.js';
 import { css, html, LitElement } from 'lit-element';
 import { Debouncer } from '@polymer/polymer/lib/utils/debounce.js';
-import { LocalizeConsistentEvaluation } from '../../lang/localize-consistent-evaluation.js';
-import { pdfExtension } from '../controllers/constants';
+import { LocalizeConsistentEvaluation } from '../../../lang/localize-consistent-evaluation.js';
+import { pdfExtension } from '../../controllers/constants';
 import { timeOut } from '@polymer/polymer/lib/utils/async.js';
 
 export class ConsistentEvaluationEvidenceFile extends LocalizeConsistentEvaluation(LitElement) {
