+++ components/left-panel/assignments/consistent-evaluation-submissions-page.js
@@ -4,12 +4,12 @@ import '@brightspace-ui/core/components/list/list.js';
 import '@brightspace-ui/core/components/colors/colors.js';
 import './consistent-evaluation-submission-item.js';
 import { css, html, LitElement } from 'lit-element/lit-element.js';
-import { tiiRefreshAction, tiiRel, tiiSubmitActionName, toggleFlagActionName, toggleIsReadActionName } from '../controllers/constants.js';
+import { tiiRefreshAction, tiiRel, tiiSubmitActionName, toggleFlagActionName, toggleIsReadActionName } from '../../controllers/constants.js';
 import { Classes } from 'd2l-hypermedia-constants';
-import { ConsistentEvalTelemetry } from '../helpers/consistent-eval-telemetry.js';
-import { convertToken } from '../helpers/converterHelpers.js';
-import { findFile } from '../helpers/submissionsAndFilesHelpers.js';
-import { LocalizeConsistentEvaluation } from '../../lang/localize-consistent-evaluation.js';
+import { ConsistentEvalTelemetry } from '../../helpers/consistent-eval-telemetry.js';
+import { convertToken } from '../../helpers/converterHelpers.js';
+import { findFile } from '../../helpers/submissionsAndFilesHelpers.js';
+import { LocalizeConsistentEvaluation } from '../../../lang/localize-consistent-evaluation.js';
 import { performSirenAction } from 'siren-sdk/src/es6/SirenAction.js';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
 import { SkeletonMixin } from '@brightspace-ui/core/components/skeleton/skeleton-mixin.js';
